Cater to the non Joe Sixpacks programmers
I did have a quick look at your Digest module and was somewhat surprised
by the following statement about add_bits:

"Most users should just ignore this method as partial bytes is very
unlikely to be of any practical use."

Actually, that statement is factually incorrect: the method is not only
practical but absolutely necessary for users who need to examine or
authenticate bit-level data in an unambiguous manner.  NIST had the
wisdom to foresee this when they wrote their specification.  These users
include cryptanalysts, protocol designers, and special-purpose
application developers who handle partial byte data.  One analyst found
a novel application of it to uniquely identify large numeric objects in
a highly-optimized multiple-precision environment.

And even if you're using the term "practical" in the sense of a value
judgment, I think it's incorrect there as well.  Admittedly, it's fair
to say that add_bits is not likely to be of practical value to the Joe
Sixpacks of the programming world.  But it would be unfair to assume
that all Digest users are Joe Sixpacks.  Programmers who work with
things like formal verification and bit-level authentication tend to be
more highly-educated and much more highly paid than your typical
programmer.  And that's very practical!  I worked with many such people
for 18 years.

It is accurate to say, though, that most users can just ignore the
add_bits method since typical applications involve only whole-byte data.